# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KESHAV",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/thermal/current/{location_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Current Thermal",
        "description": "Get current thermal stress for a location.\n\nCalculates thermal indices from the most recent weather observation.",
        "operationId": "get_current_thermal_api_v1_thermal_current__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 2026,
              "title": "Seed"
            }
          },
          {
            "name": "config_override",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Config Override"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalStressResultList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/history/{location_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Thermal History",
        "description": "Get historical thermal stress for a location over a time window.",
        "operationId": "get_thermal_history_api_v1_thermal_history__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 start (UTC)",
              "title": "Start"
            },
            "description": "ISO-8601 start (UTC)"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 end (UTC)",
              "title": "End"
            },
            "description": "ISO-8601 end (UTC)"
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 2026,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalStressResultList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/forecast/{location_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Thermal Forecast",
        "description": "Get thermal stress forecast for a location.",
        "operationId": "get_thermal_forecast_api_v1_thermal_forecast__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "horizon_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 120,
              "minimum": 1,
              "default": 48,
              "title": "Horizon Hours"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 2026,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalStressResultList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/nighttime/{location_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Nighttime Heat",
        "description": "Get nighttime heat analysis for recent nights.",
        "operationId": "get_nighttime_heat_api_v1_thermal_nighttime__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 2026,
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NighttimeHeatMetric"
                  },
                  "title": "Response Get Nighttime Heat Api V1 Thermal Nighttime  Location Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/exposure/{location_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Exposure Memory",
        "description": "Get heat exposure memory states for a location.",
        "operationId": "get_exposure_memory_api_v1_thermal_exposure__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExposureMemoryState"
                  },
                  "title": "Response Get Exposure Memory Api V1 Thermal Exposure  Location Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/daily-summary/{location_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Thermal Daily Summary",
        "description": "Get daily thermal stress summaries for a location.",
        "operationId": "get_thermal_daily_summary_api_v1_thermal_daily_summary__location_id__get",
        "parameters": [
          {
            "name": "location_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Location Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThermalDailySummary"
                  },
                  "title": "Response Get Thermal Daily Summary Api V1 Thermal Daily Summary  Location Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/calculate": {
      "post": {
        "tags": [
          "thermal"
        ],
        "summary": "Calculate Thermal",
        "description": "Calculate thermal stress from provided meteorological inputs.\n\nThis endpoint allows calculation of thermal indices from arbitrary\ntemperature, humidity, wind, radiation, etc. values without requiring\nexisting weather observations in the database.",
        "operationId": "calculate_thermal_api_v1_thermal_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThermalCalculateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalStressResultList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/provenance/{calculation_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Calculation Provenance",
        "description": "Get detailed provenance for a specific thermal calculation.",
        "operationId": "get_calculation_provenance_api_v1_thermal_provenance__calculation_id__get",
        "parameters": [
          {
            "name": "calculation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Calculation Provenance Api V1 Thermal Provenance  Calculation Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/thermal/run/{run_id}": {
      "get": {
        "tags": [
          "thermal"
        ],
        "summary": "Get Calculation Run",
        "description": "Get thermal calculation run audit log.",
        "operationId": "get_calculation_run_api_v1_thermal_run__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalCalculationRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ThermalCalculateRequest": {
        "properties": {
          "location_id": {
            "type": "string",
            "title": "Location Id"
          },
          "timestamps": {
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "type": "array",
            "title": "Timestamps"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/ThermalInput"
            },
            "type": "array",
            "title": "Inputs"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "OBSERVATION",
              "FORECAST",
              "ESTIMATED"
            ],
            "title": "Source Type",
            "default": "OBSERVATION"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThermalCalculationConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "location_id",
          "timestamps",
          "inputs"
        ],
        "title": "ThermalCalculateRequest",
        "description": "Request body for POST /api/v1/thermal/calculate."
      },
...
      "ThermalStressResultList": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ThermalStressResult"
            },
            "type": "array",
            "title": "Results"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "ThermalStressResultList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}